// src/types/interval.d.ts
export type TIntervalSchema = {
	"interval_from": number;
	"interval_to": number;
	"solid_core": number;
	"broken_core": number;
	"vien_quantity": number;
	"vien": number;
	"breccia": number;
	"cataclosite": number;
	"mylonite": number;
	"id": number;
	"cell_id": number;
};
export type TFillingFilter = "Vn" | "SCR" | "Fr" | "Tect";